Change the repository structure of the folder: 'backend'
routesディレクトリに実際のルーティング処理を記述する
- user: ユーザー登録・ログイン機能など
